+++ woodwork/statistics_utils/__init__.py
@@ -1,13 +1,13 @@
-from .get_box_plot_info_for_column import _get_box_plot_info_for_column
-from .get_describe_dict import _get_describe_dict
-from .get_histogram_values import _get_histogram_values
-from .get_mode import _get_mode
-from .get_mutual_information_dict import _get_mutual_information_dict
-from .get_numeric_value_counts_in_range import _get_numeric_value_counts_in_range
-from .get_recent_value_counts import _get_recent_value_counts
-from .get_top_values_categorical import _get_top_values_categorical
-from .get_valid_mi_columns import _get_valid_mi_columns
-from .get_value_counts import _get_value_counts
-from .infer_temporal_frequencies import _infer_temporal_frequencies
-from .make_categorical_for_mutual_info import _make_categorical_for_mutual_info
-from .replace_nans_for_mutual_info import _replace_nans_for_mutual_info
+from ._get_box_plot_info_for_column import _get_box_plot_info_for_column
+from ._get_describe_dict import _get_describe_dict
+from ._get_histogram_values import _get_histogram_values
+from ._get_mode import _get_mode
+from ._get_mutual_information_dict import _get_mutual_information_dict
+from ._get_numeric_value_counts_in_range import _get_numeric_value_counts_in_range
+from ._get_recent_value_counts import _get_recent_value_counts
+from ._get_top_values_categorical import _get_top_values_categorical
+from ._get_valid_mi_columns import _get_valid_mi_columns
+from ._get_value_counts import _get_value_counts
+from ._infer_temporal_frequencies import _infer_temporal_frequencies
+from ._make_categorical_for_mutual_info import _make_categorical_for_mutual_info
+from ._replace_nans_for_mutual_info import _replace_nans_for_mutual_info
