+++ woodwork/statistics_utils/_get_mutual_information_dict.py
@@ -2,8 +2,8 @@ from timeit import default_timer as timer
 
 from sklearn.metrics.cluster import normalized_mutual_info_score
 
-from .make_categorical_for_mutual_info import _make_categorical_for_mutual_info
-from .replace_nans_for_mutual_info import _replace_nans_for_mutual_info
+from ._make_categorical_for_mutual_info import _make_categorical_for_mutual_info
+from ._replace_nans_for_mutual_info import _replace_nans_for_mutual_info
 
 from woodwork.accessor_utils import _is_dask_dataframe, _is_koalas_dataframe
 from woodwork.utils import _update_progress, get_valid_mi_types
