+++ woodwork/statistics_utils/_replace_nans_for_mutual_info.py
@@ -1,4 +1,4 @@
-from .get_mode import _get_mode
+from ._get_mode import _get_mode
 
 from woodwork.logical_types import Double
